Refactor to custom hook
